refactor: share dataset utilities
